playback.c: select plughw, increase the frames in period
With "default" device, was unable to get audio ouput. Manaually
tested the working PCM device by using aplay and found plughw
to be working correctly. Therefore open plughw PCM device instead
of default.
Also increased the number of fames in a period by adding a define
for it as there were breaks in audio with lesser period sizes
along with frequent underruns.

Signed-off-by: Yasir Khan <yasir_electronics@yahoo.com>