voip.c: prefill the output buffer to avoid underruns
Finally found the missing piece to avoid underruns i.e. to prefill
the output buffer with arbitrary data for two periods. This gives us a
safety margin of two periods in case the producer misses an audio period.
This has resulted in elimination of underruns and their associated rattling
in output audio completely.

Signed-off-by: Yasir Khan <yasir_electronics@yahoo.com>