server.c: add audio capture and send in the single thread
Previously audio capture and transmission were being done in separate
threads. This resulted in situations where capture would be followed
by another capture instead of transmission which might result in
underruns on the receiver side. Therefore combine audio capture
and transmission in a single thread for a more deterministic behavior.

Signed-off-by: Yasir Khan <yasir_electronics@yahoo.com>